requirements.txt
pydub
requests
pyyaml
openai==0.28
tiktoken
chainlit
openai-whisper